move test to directories to match src
