+++ test/unit/data/schema/model/local/local-config.test.ts
@@ -7,19 +7,19 @@ import {instanceToPlain} from 'class-transformer';
 import {SemVer} from 'semver';
 import {beforeEach} from 'mocha';
 import os from 'os';
-import {LocalConfig} from '../../../../../src/data/schema/model/local/local-config.js';
-import {Deployment} from '../../../../../src/data/schema/model/local/deployment.js';
-import {LocalConfigSchema} from '../../../../../src/data/schema/migration/impl/local/local-config-schema.js';
-import {CTObjectMapper} from '../../../../../src/data/mapper/impl/ct-object-mapper.js';
-import {ApplicationVersions} from '../../../../../src/data/schema/model/common/application-versions.js';
+import {LocalConfig} from '../../../../../../src/data/schema/model/local/local-config.js';
+import {Deployment} from '../../../../../../src/data/schema/model/local/deployment.js';
+import {LocalConfigSchema} from '../../../../../../src/data/schema/migration/impl/local/local-config-schema.js';
+import {CTObjectMapper} from '../../../../../../src/data/mapper/impl/ct-object-mapper.js';
+import {ApplicationVersions} from '../../../../../../src/data/schema/model/common/application-versions.js';
 import {
   HEDERA_EXPLORER_VERSION,
   HEDERA_JSON_RPC_RELAY_VERSION,
   HEDERA_PLATFORM_VERSION,
   MIRROR_NODE_VERSION,
   SOLO_CHART_VERSION,
-} from '../../../../../version.js';
-import {getSoloVersion} from '../../../../../src/core/helpers.js';
+} from '../../../../../../version.js';
+import {getSoloVersion} from '../../../../../../src/core/helpers.js';
 
 describe('LocalConfig', () => {
   const schema: LocalConfigSchema = new LocalConfigSchema(new CTObjectMapper());
